refactor: restructure application folders (DTOs, Errors, Options)
