--- LensmaniaLibrary/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace LensmaniaLibrary;
-
-public class Class1
-{
-}
