+++ LensmaniaLibrary/DTOs/Auth/AuthResponse.cs
@@ -1,3 +1,3 @@
-namespace LensmaniaServer.Models;
+namespace LensmaniaLibrary.DTOs.Auth;
 
 public record AuthResponse(string Token, string Username, bool IsAdmin, bool IsPremium, bool IsActive);
