--- LensmaniaServer/Models/ApiErrorResponse.cs
@@ -1,8 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace LensmaniaServer.Models;
-
-
-public record ApiErrorResponse(
-    [property: JsonPropertyName("code")] string Code,
-    [property: JsonPropertyName("message")] string Message);
